chat - move setup components into the respective contrib
